refactor: simplify llm agent structure and remove thought manager
