+++ examples/sample_app/custom_service.py
@@ -54,7 +54,6 @@ class CustomPipeline(DefaultPipeline):
         captioner: Optional[Captioner] = None,
         punt_restorer_model: Optional[PuntRestorer] = None,
         caption_rewriter: Optional[Rewriter | BaseChatModel] = None,
-        thought_rewriter: Optional[Rewriter | BaseChatModel] = None,
         vad: Optional[VAD] = None,
         speech_enhancer: Optional[SpeechEnhancer] = None,
         speaker_encoder: Optional[SpeakerEncoder] = None,
@@ -70,7 +69,6 @@ class CustomPipeline(DefaultPipeline):
             captioner=captioner,
             punt_restorer_model=punt_restorer_model,
             caption_rewriter=caption_rewriter,
-            thought_rewriter=thought_rewriter,
             vad=vad,
             speech_enhancer=speech_enhancer,
             speaker_encoder=speaker_encoder,
