+++ src/xtalk/llm_agent/default.py
@@ -1,26 +1,26 @@
 from __future__ import annotations
 
+import asyncio
 from collections import Counter
+from contextlib import contextmanager
+from dataclasses import dataclass, field
 from datetime import datetime
 import re
-from typing import Any, AsyncIterator, Callable, Iterable, Optional
+from typing import Any, AsyncIterator, Callable, Iterable, Optional, TypeVar
 
 from langchain.chat_models.base import BaseChatModel
-from langchain_core.messages import HumanMessage, SystemMessage
+from langchain_core.messages import (
+    AIMessage,
+    BaseMessage,
+    HumanMessage,
+    SystemMessage,
+    ToolCall,
+    ToolMessage,
+)
 from langchain_core.tools import BaseTool
+from langchain_openai import ChatOpenAI
 
 from ..log_utils import logger
-from .utils.runtime import (
-    AgentSession,
-    ContextAdapter,
-    OutputPolicy,
-    PromptBuilder,
-    ScenarioSpec,
-    TurnContext,
-    get_context_data,
-)
-from .utils.interfaces import AgentContext, AgentOutput
-from .utils.template import MutableToolProvider, TemplateAgent, _format_chat_history
 from .tools import (
     build_set_emotion_tool,
     build_set_speed_tool,
@@ -30,6 +30,8 @@ from .tools import (
     build_web_search_tool,
 )
 from .tools.retrievers import LOCAL_SEARCH_TOOL, build_local_search_tool
+from .tools.utils import build_tool_call_result_payload
+from .interfaces import Agent, AgentContext, AgentOutput
 
 _EMBEDDING_PROCESSING_PROMPT = """
 You are a helpful assistant whose only task is to generate a short, natural-sounding transitional sentence to indicate that you are aware of a Doc uploaded.
@@ -45,259 +47,195 @@ You should mention the Doc summary.
 Your response should be catered to the given Chat history, e.g. respond in the same language as the User.
 """
 
+_AGENT_CONTEXTS_KEY = "_agent_runtime_contexts"
+T = TypeVar("T")
 
-class DefaultContextAdapter(ContextAdapter):
-    """Adapt accepted agent context updates into the default ``TurnContext``."""
 
-    def adapt(
+@dataclass
+class AgentSession:
+    """Mutable per-session state for the default LLM agent.
+
+    Parameters
+    ----------
+    messages : list[BaseMessage]
+        Conversation history, including system, user, assistant, and tool
+        messages.
+    metadata : dict[str, Any]
+        Session-scoped mutable state used by context updates and dynamic tools.
+    """
+
+    messages: list[BaseMessage] = field(default_factory=list)
+    metadata: dict[str, Any] = field(default_factory=dict)
+
+
+@dataclass
+class _TurnState:
+    """Internal turn-scoped state derived from accepted contexts."""
+
+    speaker_id: str | None = None
+    caption: str | None = None
+
+
+def get_context_data(
+    session: AgentSession,
+    context_type: str,
+) -> dict[str, Any]:
+    """Return stored agent-context data for one logical context type.
+
+    Parameters
+    ----------
+    session : AgentSession
+        Mutable runtime session state.
+    context_type : str
+        Logical context stream name such as ``"caption"``.
+
+    Returns
+    -------
+    dict[str, Any]
+        Stored payload for the context type, or an empty dict.
+    """
+
+    raw_contexts = session.metadata.get(_AGENT_CONTEXTS_KEY)
+    if not isinstance(raw_contexts, dict):
+        return {}
+    data = raw_contexts.get(context_type)
+    if not isinstance(data, dict):
+        return {}
+    return dict(data)
+
+
+def _format_chat_history(
+    messages: list[BaseMessage],
+    *,
+    with_system: bool = False,
+) -> str | None:
+    """Render plain-text chat history from LangChain messages."""
+
+    if not messages:
+        return None
+    lines: list[str] = []
+    for message in messages:
+        role = "System"
+        if isinstance(message, HumanMessage):
+            role = "User"
+        elif isinstance(message, AIMessage):
+            role = "Assistant"
+        if role == "System" and not with_system:
+            continue
+        content = message.content
+        lines.append(f"{role}: {content if isinstance(content, str) else str(content)}")
+    return "\n".join(lines)
+
+
+class MutableToolProvider:
+    """Provide a clone-safe list of tools that can be extended at runtime.
+
+    Parameters
+    ----------
+    tools : list[BaseTool | Callable[[], BaseTool]] | None, optional
+        Initial tool instances or factories.
+    """
+
+    def __init__(
         self,
-        session: AgentSession,
-        request: dict[str, Any],
-    ) -> TurnContext:
-        """Adapt the current session-scoped agent context.
+        tools: Optional[list[BaseTool | Callable[[], BaseTool]]] = None,
+    ) -> None:
+        self._tool_factories = self.normalize_tool_specs(tools or [])
+
+    def add_tools(
+        self,
+        tools: list[BaseTool | Callable[[], BaseTool]],
+    ) -> None:
+        """Append tools or tool factories.
+
+        Parameters
+        ----------
+        tools : list[BaseTool | Callable[[], BaseTool]]
+            Tool instances or factories to append.
+        """
+
+        self._tool_factories.extend(self.normalize_tool_specs(tools))
+
+    def get_tool_specs(self) -> list[Callable[[], BaseTool]]:
+        """Return clone-safe tool factories.
+
+        Returns
+        -------
+        list[Callable[[], BaseTool]]
+            Normalized factories used by this provider.
+        """
+
+        return list(self._tool_factories)
+
+    def get_tools(self, session: AgentSession) -> list[BaseTool]:
+        """Return session-scoped tool instances for the current session.
 
         Parameters
         ----------
         session : AgentSession
-            Mutable runtime session state.
-        request : dict[str, Any]
-            Current turn input. Unused by the default adapter.
+            Runtime session state used to cache tool instances.
 
         Returns
         -------
-        TurnContext
-            Stable turn context for the runtime.
+        list[BaseTool]
+            Tool instances reused within the current session.
         """
 
-        del request
-        # TODO: Remove the default agent's reliance on context-derived speaker,
-        # caption, and thought state once scenario-native state is introduced
-        # for these features.
-        speaker_context = get_context_data(session, "speaker")
-        caption_context = get_context_data(session, "caption")
-        thought_context = get_context_data(session, "thought")
-        return TurnContext(
-            speaker_id=speaker_context.get("speaker_id") or None,
-            caption=caption_context.get("text") or None,
-            thought=thought_context.get("text") or None,
-        )
+        cache_key = f"_mutable_tool_provider_cache_{id(self)}"
+        cached_tools = session.metadata.get(cache_key)
+        if isinstance(cached_tools, list) and all(
+            isinstance(tool, BaseTool) for tool in cached_tools
+        ):
+            return cached_tools
 
+        tools: list[BaseTool] = []
+        for factory in self._tool_factories:
+            try:
+                tool = factory()
+            except Exception as exc:
+                logger.warning("Tool factory %r failed: %s", factory, exc)
+                continue
+            if isinstance(tool, BaseTool):
+                tools.append(tool)
+            else:
+                logger.warning(
+                    "Tool factory %r returned non-BaseTool: %s",
+                    factory,
+                    type(tool),
+                )
+        session.metadata[cache_key] = tools
+        return tools
 
-class DefaultPromptBuilder(PromptBuilder):
-    """Build the default scenario prompt and user message."""
-
-    BASE_PROMPT: str = """
-You are a friendly conversational partner whose response will be converted to speech using TTS. Please follow rules below:
-1. Respond with the same language as user.
-Examples:
-- user: 你好。
-- assistant: 你好呀，今天感觉怎么样？
-- user: Hello.
-- assistant: Hello, how are you today?
-
-2. Your response should not contain content that cannot be synthesize by the TTS model, such as parentheses, ordered lists (starting by - ), etc. Numbers should be written in English words rather than Arabic numerals.
-
-3. Your response should be informative and adequately detailed, but avoid unnecessary repetition or filler. Keep it suitable for spoken delivery.
-
-4. If you find user input (ASR result) unclear, incomplete, or likely incorrect — for example:
-- contains obvious ASR hallucinations,
-- contains broken words or meaningless fragments,
-- does not form a valid sentence,
-- semantic intention cannot be determined,
-then DO NOT guess the user's meaning.
-Instead, politely ask the user to repeat their last utterance.
-
-5. Each distinct speaker ID corresponds to a separate dialogue user.
-The system should distinguish users based on their speaker IDs, with one user mapped to one speaker ID.
-
-6. You have access to tools. You MUST use them proactively:
-- get_time: call when user asks about current time, date, or day of week.
-- web_search: you MUST default to searching for ANY question about specific facts, including but not limited to:
-  * Weather, news, current events, real-time data (stock prices, sports scores, exchange rates)
-  * Specific places, buildings, campuses, addresses, floor numbers, room numbers, opening hours
-  * Restaurants, shops, cafes, businesses and their details (location, menu, price, how many)
-  * Specific people, organizations, companies, products, events
-  * Questions involving numbers, statistics, rankings, or comparisons that require accuracy
-  * Any question where giving an INCORRECT answer is worse than taking a moment to search
-- set_voice: call when user asks to change voice or sound like someone.
-- set_speed: call when user asks to speak faster or slower.
-- GOLDEN RULE: If you are not 100% certain your answer is accurate AND up-to-date, call web_search. When in doubt, ALWAYS search.
-- NEVER say "I cannot access real-time information" or "I don't have internet access". You have search tools — USE THEM.
-- NEVER answer specific factual questions from memory alone — search first, then answer based on search results.
-
-7. When citing times, numbers, names, or other specific facts from search results, you SHOULD reproduce them faithfully. Do NOT reinterpret or convert values based on your assumptions. For example, if search results say "10:30", treat it as 10:30 AM unless the source explicitly says PM or evening.
-
-8. SEARCH QUERY RULE: When constructing a web_search query, ALWAYS replace relative time references ("今天", "昨天", "明天", "上个月", "去年", "today", "yesterday", etc.) with the actual date from <current_date>. For example, if today is 2026-02-28 and the user asks "今天NBA有哪些比赛", your query should be "2026年2月28日 NBA比赛赛程", NOT "今天NBA有哪些比赛".
-
-你是一位友好的对话伙伴，你的回复会通过 TTS 转成语音。请遵守以下规则：
-
-1. 用和用户相同的语言回复。
-示例：
-- user: 你好。
-- assistant: 你好呀，今天感觉怎么样？
-- user: Hello.
-- assistant: Hello, how are you today?
-
-2. 你的回复中不能出现 TTS 无法合成的内容，例如括号、编号列表（以- 开始）等。数字要用英文单词书写，不要使用阿拉伯数字。
-
-3. 你的回复应当信息充分、适当详细，但避免不必要的重复或废话。回复长度要适合语音播报。
-
-4. 如果你发现用户输入（ASR 结果）不清晰、不完整或可能有误，例如：
-- 包含明显的 ASR 幻觉内容；
-- 包含残缺的词语或无意义的片段；
-- 无法构成有效句子；
-- 无法判断其语义意图；
-那么不要猜测用户的意思。
-请礼貌地请求用户重复上一句内容。
-5. 有几个不同说话人id就有几个不同的对话用户，每个说话人id对应一个用户，你要根据说话人id来区分用户。
-
-6. 你可以使用工具，必须主动调用：
-- get_time：用户问当前时间、日期、星期几时调用。
-- web_search：遇到任何关于具体事实的问题时，必须优先搜索，包括但不限于：
-  * 天气、新闻、时事、实时数据（股价、比分、汇率等）
-  * 具体地点、建筑、校园、地址、楼层、房间号、营业时间
-  * 餐厅、商店、咖啡厅、商家及其详细信息（位置、菜单、价格、数量）
-  * 具体人物、机构、公司、产品、事件
-  * 涉及数字、统计、排名或需要准确性的比较类问题
-  * 任何回答错误比多花一点时间搜索更糟糕的问题
-- set_voice：用户要求换声音或模仿某人声音时调用。
-- set_speed：用户要求说快一点或慢一点时调用。
-- 黄金原则：如果你不能百分之百确定答案准确且是最新的，就调用 web_search。有疑问时，永远先搜索。
-- 绝对不要说"我无法获取实时信息"或"我没有联网能力"。你拥有搜索工具，请使用它们。
-- 绝对不要仅凭记忆回答具体的事实性问题——先搜索，再根据搜索结果回答。
-
-7. 引用搜索结果中的时间、数字、名称等具体事实时，应该忠实于原文，不要根据自己的推测重新解读。例如搜索结果写"10:30"，应说"上午十点三十分"，除非原文明确标注是下午或晚上。
-
-8. 搜索用语规则：构造 web_search 的 query 时，必须将"今天"、"昨天"、"明天"、"上个月"、"去年"等相对时间词替换为 <current_date> 中的具体日期。例如今天是2026-02-28，用户问"今天NBA有哪些比赛"，你的 query 应为"2026年2月28日 NBA比赛赛程"，而不是"今天NBA有哪些比赛"。
-"""
-
-    CONTEXT_AWARE_PROMPT: str = (
-        """
-You are a multimodal conversational assistant with access to:
-1) Non-verbal environmental context extracted from recent audio, wrapped in <caption>...</caption>.
-2) Your internal reasoning summary for the latest turn, wrapped in <thought>...</thought>.
-
-About <caption>:
-- It describes the user's environment, emotional cues, ambient sounds, and relevant non-verbal context.
-- It may contain incomplete or approximate descriptions; treat it as helpful hints, not absolute truth.
-- Use it only to enrich understanding and respond more naturally, not to hallucinate details that are not implied.
-- DO NOT reveal <caption> content directly in your replies.
-
-About <thought>:
-- It summarizes your internal intention and reasoning for this turn.
-- It represents your state before generating the answer.
-- DO NOT continue thinking while composing the answer.
-- DO NOT reveal the content of <thought> or mention that you have internal thoughts.
-- Treat it as internal context to adjust tone, structure, and direction of your reply.
-
-When generating your final response:
-- Use both <caption> and <thought> as private hints to better understand the user's situation.
-- Never output the tags themselves, nor refer to them explicitly.
-- Do NOT invent nonexistent sensations, emotions, or events.
-- Focus on giving a helpful, grounded, natural reply to the user's last message.
-- If caption and user text conflict, ALWAYS prioritize the user's explicit message.
-
-Caption and thought:
-""".strip()
-    )
-
-    def __init__(self, system_prompt: str = BASE_PROMPT) -> None:
-        """Initialize the default prompt builder.
+    @staticmethod
+    def normalize_tool_specs(
+        tools: list[BaseTool | Callable[[], BaseTool]],
+    ) -> list[Callable[[], BaseTool]]:
+        """Normalize tools into zero-argument factories.
 
         Parameters
         ----------
-        system_prompt : str, optional
-            Base system prompt for the scenario.
-        """
-
-        self.system_prompt = system_prompt
-        self.session_system_prompt = f"{system_prompt}\n\n{self.CONTEXT_AWARE_PROMPT}"
-
-    def build_system_prompt(
-        self,
-        session: AgentSession,
-        turn_context: TurnContext,
-    ) -> str:
-        """Build the system prompt for the current turn.
-
-        Parameters
-        ----------
-        session : AgentSession
-            Runtime session state.
-        turn_context : TurnContext
-            Structured turn context.
+        tools : list[BaseTool | Callable[[], BaseTool]]
+            Tool instances or factories.
 
         Returns
         -------
-        str
-            Complete system prompt for the model.
+        list[Callable[[], BaseTool]]
+            Factory list.
         """
 
-        del session
-        parts = [
-            f"<current_date>{datetime.now().strftime('%Y-%m-%d %A')}</current_date>"
-        ]
-        # TODO: Move caption/thought prompt decoration behind a dedicated state
-        # model instead of depending on context-derived turn fields.
-        if turn_context.caption:
-            parts.append(f"<caption>{turn_context.caption.strip()}</caption>")
-        if turn_context.thought:
-            parts.append(f"<thought>{turn_context.thought.strip()}</thought>")
-        return f"{self.session_system_prompt}\n" + "\n".join(parts)
-
-    def build_user_message(
-        self,
-        request: dict[str, Any],
-        turn_context: TurnContext,
-    ) -> str:
-        """Build the user message content.
-
-        Parameters
-        ----------
-        request : dict[str, Any]
-            Turn input.
-        turn_context : TurnContext
-            Structured turn context.
-
-        Returns
-        -------
-        str
-            User message content passed to the model.
-        """
-
-        content = str(request.get("content", ""))
-        if turn_context.speaker_id:
-            return (
-                f"The current speaker is {turn_context.speaker_id}, saying: {content}"
-            )
-        return content
-
-
-class TTSOutputPolicy(OutputPolicy):
-    """Normalize assistant text for speech synthesis."""
-
-    def filter_text(self, text: str) -> str:
-        """Remove unsupported formatting from a response fragment.
-
-        Parameters
-        ----------
-        text : str
-            Raw response fragment.
-
-        Returns
-        -------
-        str
-            TTS-friendly text.
-        """
-
-        filtered = (
-            text.replace("#", "").replace("**", "").replace("`", "").replace("-", "")
-        )
-        return re.sub(r"(\d+)\.", r"\1", filtered)
+        factories: list[Callable[[], BaseTool]] = []
+        for tool in tools:
+            if isinstance(tool, BaseTool):
+                factories.append(lambda tool=tool: tool)
+            elif callable(tool):
+                factories.append(tool)
+            else:
+                logger.warning("Unsupported tool spec: %r", tool)
+        return factories
 
 
 class DefaultToolProvider(MutableToolProvider):
-    """Build the default tool set for each turn."""
+    """Build the default tool set for each session."""
 
     def __init__(
         self,
@@ -325,27 +263,20 @@ class DefaultToolProvider(MutableToolProvider):
             self._build_default_tool_factories() if tools is None else tools
         )
 
-    def get_tools(
-        self,
-        session: AgentSession,
-        turn_context: TurnContext,
-    ) -> list[BaseTool]:
-        """Build the tool set for the current turn.
+    def get_tools(self, session: AgentSession) -> list[BaseTool]:
+        """Build the tool set for the current session.
 
         Parameters
         ----------
         session : AgentSession
             Runtime session state.
-        turn_context : TurnContext
-            Structured turn context.
 
         Returns
         -------
         list[BaseTool]
-            Enabled tools for this turn.
+            Enabled tools for this session.
         """
 
-        del turn_context
         tools_by_name: dict[str, BaseTool] = {}
         for factory in self._tool_factories:
             try:
@@ -362,9 +293,7 @@ class DefaultToolProvider(MutableToolProvider):
                 continue
             tools_by_name[tool.name] = tool
 
-        for name, dynamic_tool in (
-            session.metadata.get("dynamic_tools", {}) or {}
-        ).items():
+        for name, dynamic_tool in (session.metadata.get("dynamic_tools", {}) or {}).items():
             if isinstance(dynamic_tool, BaseTool):
                 tools_by_name[name] = dynamic_tool
 
@@ -464,60 +393,120 @@ def summarize_embedding_doc(doc: str, max_sentences: int | None = None) -> str:
     return " ".join(sentences[idx] for idx in selected)
 
 
-def build_default_scenario(
-    *,
-    model: BaseChatModel,
-    system_prompt: str = DefaultPromptBuilder.BASE_PROMPT,
-    voice_names: Optional[list[str]] = None,
-    emotions: Optional[list[str]] = None,
-    tools: Optional[list[BaseTool | Callable[[], BaseTool]]] = None,
-) -> tuple[ScenarioSpec, DefaultPromptBuilder, DefaultToolProvider]:
-    """Assemble the default scenario components.
+class DefaultAgent(Agent):
+    """Default speech-first conversational agent implementation."""
 
-    Parameters
-    ----------
-    model : BaseChatModel
-        Chat model used by the runtime.
-    system_prompt : str, optional
-        Base system prompt.
-    voice_names : list[str] | None, optional
-        Available voice names.
-    emotions : list[str] | None, optional
-        Available emotion names.
-    tools : list[BaseTool | Callable[[], BaseTool]] | None, optional
-        Explicit tool set or factories.
+    BASE_PROMPT: str = """
+You are a friendly conversational partner whose response will be converted to speech using TTS. Please follow rules below:
+1. Respond with the same language as user.
+Examples:
+- user: 你好。
+- assistant: 你好呀，今天感觉怎么样？
+- user: Hello.
+- assistant: Hello, how are you today?
 
-    Returns
-    -------
-    tuple[ScenarioSpec, DefaultPromptBuilder, DefaultToolProvider]
-        Scenario spec plus the reusable default prompt builder and tool provider.
-    """
+2. Your response should not contain content that cannot be synthesize by the TTS model, such as parentheses, ordered lists (starting by - ), etc. Numbers should be written in English words rather than Arabic numerals.
 
-    prompt_builder = DefaultPromptBuilder(system_prompt=system_prompt)
-    output_policy = TTSOutputPolicy()
-    tool_provider = DefaultToolProvider(
-        voice_names=voice_names,
-        emotions=emotions,
-        tools=tools,
+3. Your response should be informative and adequately detailed, but avoid unnecessary repetition or filler. Keep it suitable for spoken delivery.
+
+4. If you find user input (ASR result) unclear, incomplete, or likely incorrect — for example:
+- contains obvious ASR hallucinations,
+- contains broken words or meaningless fragments,
+- does not form a valid sentence,
+- semantic intention cannot be determined,
+then DO NOT guess the user's meaning.
+Instead, politely ask the user to repeat their last utterance.
+
+5. Each distinct speaker ID corresponds to a separate dialogue user.
+The system should distinguish users based on their speaker IDs, with one user mapped to one speaker ID.
+
+6. You have access to tools. You MUST use them proactively:
+- get_time: call when user asks about current time, date, or day of week.
+- web_search: you MUST default to searching for ANY question about specific facts, including but not limited to:
+  * Weather, news, current events, real-time data (stock prices, sports scores, exchange rates)
+  * Specific places, buildings, campuses, addresses, floor numbers, room numbers, opening hours
+  * Restaurants, shops, cafes, businesses and their details (location, menu, price, how many)
+  * Specific people, organizations, companies, products, events
+  * Questions involving numbers, statistics, rankings, or comparisons that require accuracy
+  * Any question where giving an INCORRECT answer is worse than taking a moment to search
+- set_voice: call when user asks to change voice or sound like someone.
+- set_speed: call when user asks to speak faster or slower.
+- GOLDEN RULE: If you are not 100% certain your answer is accurate AND up-to-date, call web_search. When in doubt, ALWAYS search.
+- NEVER say "I cannot access real-time information" or "I don't have internet access". You have search tools — USE THEM.
+- NEVER answer specific factual questions from memory alone — search first, then answer based on search results.
+
+7. When citing times, numbers, names, or other specific facts from search results, you SHOULD reproduce them faithfully. Do NOT reinterpret or convert values based on your assumptions. For example, if search results say "10:30", treat it as 10:30 AM unless the source explicitly says PM or evening.
+
+8. SEARCH QUERY RULE: When constructing a web_search query, ALWAYS replace relative time references ("今天", "昨天", "明天", "上个月", "去年", "today", "yesterday", etc.) with the actual date from <current_date>. For example, if today is 2026-02-28 and the user asks "今天NBA有哪些比赛", your query should be "2026年2月28日 NBA比赛赛程", NOT "今天NBA有哪些比赛".
+
+你是一位友好的对话伙伴，你的回复会通过 TTS 转成语音。请遵守以下规则：
+
+1. 用和用户相同的语言回复。
+示例：
+- user: 你好。
+- assistant: 你好呀，今天感觉怎么样？
+- user: Hello.
+- assistant: Hello, how are you today?
+
+2. 你的回复中不能出现 TTS 无法合成的内容，例如括号、编号列表（以- 开始）等。数字要用英文单词书写，不要使用阿拉伯数字。
+
+3. 你的回复应当信息充分、适当详细，但避免不必要的重复或废话。回复长度要适合语音播报。
+
+4. 如果你发现用户输入（ASR 结果）不清晰、不完整或可能有误，例如：
+- 包含明显的 ASR 幻觉内容；
+- 包含残缺的词语或无意义的片段；
+- 无法构成有效句子；
+- 无法判断其语义意图；
+那么不要猜测用户的意思。
+请礼貌地请求用户重复上一句内容。
+5. 有几个不同说话人id就有几个不同的对话用户，每个说话人id对应一个用户，你要根据说话人id来区分用户。
+
+6. 你可以使用工具，必须主动调用：
+- get_time：用户问当前时间、日期、星期几时调用。
+- web_search：遇到任何关于具体事实的问题时，必须优先搜索，包括但不限于：
+  * 天气、新闻、时事、实时数据（股价、比分、汇率等）
+  * 具体地点、建筑、校园、地址、楼层、房间号、营业时间
+  * 餐厅、商店、咖啡厅、商家及其详细信息（位置、菜单、价格、数量）
+  * 具体人物、机构、公司、产品、事件
+  * 涉及数字、统计、排名或需要准确性的比较类问题
+  * 任何回答错误比多花一点时间搜索更糟糕的问题
+- set_voice：用户要求换声音或模仿某人声音时调用。
+- set_speed：用户要求说快一点或慢一点时调用。
+- 黄金原则：如果你不能百分之百确定答案准确且是最新的，就调用 web_search。有疑问时，永远先搜索。
+- 绝对不要说"我无法获取实时信息"或"我没有联网能力"。你拥有搜索工具，请使用它们。
+- 绝对不要仅凭记忆回答具体的事实性问题——先搜索，再根据搜索结果回答。
+
+7. 引用搜索结果中的时间、数字、名称等具体事实时，应该忠实于原文，不要根据自己的推测重新解读。例如搜索结果写"10:30"，应说"上午十点三十分"，除非原文明确标注是下午或晚上。
+
+8. 搜索用语规则：构造 web_search 的 query 时，必须将"今天"、"昨天"、"明天"、"上个月"、"去年"等相对时间词替换为 <current_date> 中的具体日期。例如今天是2026-02-28，用户问"今天NBA有哪些比赛"，你的 query 应为"2026年2月28日 NBA比赛赛程"，而不是"今天NBA有哪些比赛"。
+"""
+
+    CONTEXT_AWARE_PROMPT: str = (
+        """
+You are a multimodal conversational assistant with access to:
+1) Non-verbal environmental context extracted from recent audio, wrapped in <caption>...</caption>.
+
+About <caption>:
+- It describes the user's environment, emotional cues, ambient sounds, and relevant non-verbal context.
+- It may contain incomplete or approximate descriptions; treat it as helpful hints, not absolute truth.
+- Use it only to enrich understanding and respond more naturally, not to hallucinate details that are not implied.
+- DO NOT reveal <caption> content directly in your replies.
+
+When generating your final response:
+- Use <caption> as a private hint to better understand the user's situation.
+- Never output the tags themselves, nor refer to them explicitly.
+- Do NOT invent nonexistent sensations, emotions, or events.
+- Focus on giving a helpful, grounded, natural reply to the user's last message.
+- If caption and user text conflict, ALWAYS prioritize the user's explicit message.
+
+Caption:
+""".strip()
     )
-    scenario = ScenarioSpec(
-        name="default",
-        context_adapter=DefaultContextAdapter(),
-        prompt_builder=prompt_builder,
-        tool_provider=tool_provider,
-        output_policy=output_policy,
-        hooks=[],
-    )
-    return scenario, prompt_builder, tool_provider
-
-
-class DefaultAgent(TemplateAgent):
-    """Default scenario agent built on top of ``TemplateAgent``."""
 
     def __init__(
         self,
         model: BaseChatModel | dict[str, Any],
-        system_prompt: str = DefaultPromptBuilder.BASE_PROMPT,
+        system_prompt: str = BASE_PROMPT,
         voice_names: Optional[list[str]] = None,
         emotions: Optional[list[str]] = None,
         tools: Optional[list[BaseTool | Callable[[], BaseTool]]] = None,
@@ -538,28 +527,298 @@ class DefaultAgent(TemplateAgent):
             Explicit tool set or factories.
         """
 
-        resolved_model = TemplateAgent.coerce_model(model)
-        normalized_voice_names = list(voice_names or [])
-        normalized_emotions = list(emotions or [])
-        scenario, _, tool_provider = build_default_scenario(
-            model=resolved_model,
-            system_prompt=system_prompt,
-            voice_names=normalized_voice_names,
-            emotions=normalized_emotions,
+        self._model = self._coerce_model(model)
+        self.system_prompt = system_prompt
+        self.voice_names = list(voice_names or [])
+        self.emotions = list(emotions or [])
+        self._tool_provider = DefaultToolProvider(
+            voice_names=self.voice_names,
+            emotions=self.emotions,
             tools=tools,
         )
-        self.voice_names = normalized_voice_names
-        self.emotions = normalized_emotions
-        super().__init__(
-            model=resolved_model,
-            scenario=scenario,
-            clone_kwargs={
-                "system_prompt": system_prompt,
-                "voice_names": normalized_voice_names,
-                "emotions": normalized_emotions,
-            },
-            tool_provider=tool_provider,
+        self._session = AgentSession()
+
+    @staticmethod
+    def _coerce_model(model: BaseChatModel | dict[str, Any]) -> BaseChatModel:
+        """Coerce model configuration into a concrete chat model.
+
+        Parameters
+        ----------
+        model : BaseChatModel | dict[str, Any]
+            Chat model instance or ``ChatOpenAI`` configuration dict.
+
+        Returns
+        -------
+        BaseChatModel
+            Concrete chat model instance.
+        """
+
+        if isinstance(model, dict):
+            return ChatOpenAI(**model)
+        return model
+
+    @property
+    def model(self) -> BaseChatModel:
+        """Return the backing model."""
+
+        return self._model
+
+    @model.setter
+    def model(self, model: BaseChatModel) -> None:
+        """Update the backing model.
+
+        Parameters
+        ----------
+        model : BaseChatModel
+            New backing model.
+        """
+
+        self._model = model
+
+    @property
+    def session_history(self) -> list[BaseMessage]:
+        """Expose session history for compatibility."""
+
+        return self._session.messages
+
+    @session_history.setter
+    def session_history(self, messages: list[BaseMessage]) -> None:
+        """Replace session history for compatibility.
+
+        Parameters
+        ----------
+        messages : list[BaseMessage]
+            New session message list.
+        """
+
+        self._session.messages = messages
+
+    @contextmanager
+    def _temporary_event_loop(self) -> Iterable[asyncio.AbstractEventLoop]:
+        """Create a temporary event loop and clean it up on exit."""
+
+        loop = asyncio.new_event_loop()
+        try:
+            yield loop
+        finally:
+            try:
+                loop.run_until_complete(loop.shutdown_asyncgens())
+            except Exception:
+                pass
+            try:
+                loop.run_until_complete(loop.shutdown_default_executor())
+            except Exception:
+                pass
+            loop.close()
+
+    def _sync_iter_from_async(self, async_iter: AsyncIterator[T]) -> Iterable[T]:
+        """Convert an async iterator into a synchronous generator."""
+
+        with self._temporary_event_loop() as loop:
+            try:
+                while True:
+                    try:
+                        item = loop.run_until_complete(async_iter.__anext__())
+                    except StopAsyncIteration:
+                        break
+                    yield item
+            finally:
+                aclose = getattr(async_iter, "aclose", None)
+                if callable(aclose):
+                    try:
+                        loop.run_until_complete(aclose())
+                    except Exception:
+                        pass
+
+    def _filter_text(self, text: str) -> str:
+        """Normalize one response fragment for TTS output."""
+
+        filtered = text.replace("#", "").replace("**", "").replace("`", "").replace(
+            "-",
+            "",
         )
+        return re.sub(r"(\d+)\.", r"\1", filtered)
+
+    def _build_turn_state(self) -> _TurnState:
+        """Build internal turn state from accepted context streams."""
+
+        speaker_context = get_context_data(self._session, "speaker")
+        caption_context = get_context_data(self._session, "caption")
+        return _TurnState(
+            speaker_id=speaker_context.get("speaker_id") or None,
+            caption=caption_context.get("text") or None,
+        )
+
+    def _build_system_prompt(self, turn_state: _TurnState) -> str:
+        """Build the system prompt for the current turn."""
+
+        parts = [f"<current_date>{datetime.now().strftime('%Y-%m-%d %A')}</current_date>"]
+        if turn_state.caption:
+            parts.append(f"<caption>{turn_state.caption.strip()}</caption>")
+        session_system_prompt = f"{self.system_prompt}\n\n{self.CONTEXT_AWARE_PROMPT}"
+        return f"{session_system_prompt}\n" + "\n".join(parts)
+
+    def _build_user_message(
+        self,
+        request: dict[str, Any],
+        turn_state: _TurnState,
+    ) -> str:
+        """Build the user message content passed to the model."""
+
+        content = str(request.get("content", ""))
+        if turn_state.speaker_id:
+            return f"The current speaker is {turn_state.speaker_id}, saying: {content}"
+        return content
+
+    def _set_system_prompt(self, turn_state: _TurnState) -> None:
+        """Ensure the first session message contains the current system prompt."""
+
+        prompt = self._build_system_prompt(turn_state)
+        if self._session.messages and isinstance(self._session.messages[0], SystemMessage):
+            self._session.messages[0].content = prompt
+            return
+        self._session.messages.insert(0, SystemMessage(content=prompt))
+
+    def _accept_context_update(self, context: AgentContext) -> None:
+        """Merge an incremental context update into the session state."""
+
+        context_type = str(context.get("type", "") or "").strip()
+        if not context_type:
+            return
+        payload = context.get("data") or {}
+        if not isinstance(payload, dict):
+            return
+
+        raw_contexts = self._session.metadata.setdefault(_AGENT_CONTEXTS_KEY, {})
+        if not isinstance(raw_contexts, dict):
+            raw_contexts = {}
+            self._session.metadata[_AGENT_CONTEXTS_KEY] = raw_contexts
+
+        existing = raw_contexts.get(context_type)
+        if not isinstance(existing, dict):
+            existing = {}
+
+        merged = dict(existing)
+        merged.update(payload)
+        raw_contexts[context_type] = merged
+
+    def _build_runtime_request(self, context: AgentContext) -> dict[str, Any] | None:
+        """Build an internal runtime request from one accepted context update."""
+
+        context_type = str(context.get("type", "") or "")
+        if context_type != "asr_final":
+            return None
+
+        payload = context.get("data") or {}
+        if not isinstance(payload, dict):
+            return None
+        return {"content": str(payload.get("text", ""))}
+
+    @staticmethod
+    def _normalize_tool_call(tool_call: ToolCall | dict[str, Any]) -> ToolCall:
+        """Convert LangChain tool-call payloads into a stable dict form."""
+
+        if isinstance(tool_call, dict):
+            name = str(tool_call.get("name") or tool_call.get("tool") or "")
+            args = tool_call.get("args") or tool_call.get("arguments") or {}
+            call_id = tool_call.get("id") or tool_call.get("tool_call_id")
+            if isinstance(args, dict):
+                return ToolCall(name=name, args=args, id=call_id)
+        return tool_call
+
+    async def _invoke_tool(
+        self,
+        *,
+        tools_map: dict[str, BaseTool],
+        tool_call: ToolCall,
+    ) -> ToolMessage:
+        """Invoke one tool call and normalize the result as ``ToolMessage``."""
+
+        name = tool_call["name"]
+        tool = tools_map.get(name)
+        if tool is None:
+            return ToolMessage(
+                content=f"Tool {name} not found",
+                tool_call_id=tool_call["id"],
+            )
+        tool_input = tool_call.get("args", {})
+        try:
+            if hasattr(tool, "ainvoke"):
+                result = await tool.ainvoke(tool_input)
+            else:
+                loop = asyncio.get_running_loop()
+                result = await loop.run_in_executor(None, tool.invoke, tool_input)
+            if isinstance(result, ToolMessage):
+                return result
+            return ToolMessage(
+                content=str(result),
+                tool_call_id=tool_call["id"],
+            )
+        except Exception as exc:
+            logger.warning("Tool invocation failed for %s: %s", name, exc)
+            return ToolMessage(
+                content=f"Tool {name} invocation failed: {exc}",
+                tool_call_id=tool_call["id"],
+            )
+
+    async def _stream_request(
+        self,
+        request: dict[str, Any],
+    ) -> AsyncIterator[AgentOutput]:
+        """Stream response items for one runtime request."""
+
+        turn_state = self._build_turn_state()
+        self._set_system_prompt(turn_state)
+
+        tools = self._tool_provider.get_tools(self._session)
+        model_with_tools = self.model.bind_tools(tools) if tools else self.model
+        tools_map = {tool.name: tool for tool in tools}
+
+        user_content = self._build_user_message(request, turn_state)
+        if turn_state.speaker_id:
+            user_message = HumanMessage(content=user_content, name=turn_state.speaker_id)
+        else:
+            user_message = HumanMessage(content=user_content)
+        self._session.messages.append(user_message)
+
+        while True:
+            response_message = AIMessage(content="")
+            self._session.messages.append(response_message)
+
+            gathered = None
+            async for chunk in model_with_tools.astream(self._session.messages):
+                text = self._filter_text(str(chunk.content or ""))
+                if text:
+                    response_message.content += text
+                    yield text
+                gathered = chunk if gathered is None else gathered + chunk
+
+            tool_calls = (
+                list(getattr(gathered, "tool_calls", []) or []) if gathered else []
+            )
+            if tool_calls:
+                response_message.tool_calls = tool_calls
+            if not tool_calls:
+                return
+
+            for tool_call in tool_calls:
+                normalized_tool_call = self._normalize_tool_call(tool_call)
+                yield ToolCall(
+                    name=normalized_tool_call["name"],
+                    args=dict(normalized_tool_call["args"]),
+                    id=normalized_tool_call["id"],
+                )
+                tool_result = await self._invoke_tool(
+                    tools_map=tools_map,
+                    tool_call=normalized_tool_call,
+                )
+                self._session.messages.append(tool_result)
+                result_content = str(getattr(tool_result, "content", ""))
+                yield build_tool_call_result_payload(
+                    name=normalized_tool_call["name"],
+                    args=dict(normalized_tool_call["args"]),
+                    content=result_content,
+                )
 
     def accept(self, context: AgentContext) -> Iterable[AgentOutput]:
         """Accept a context update and return any triggered stream items.
@@ -571,7 +830,7 @@ class DefaultAgent(TemplateAgent):
 
         Returns
         -------
-        Iterable[AgentStreamItem]
+        Iterable[AgentOutput]
             Streamed response items triggered by the update.
         """
 
@@ -590,11 +849,11 @@ class DefaultAgent(TemplateAgent):
 
         Yields
         ------
-        AgentStreamItem
+        AgentOutput
             Streamed response items triggered by the update.
         """
 
-        self.runtime.accept(context)
+        self._accept_context_update(context)
         context_type = str(context.get("type", "") or "")
         if context_type == "embedding":
             payload = context.get("data") or {}
@@ -606,9 +865,7 @@ class DefaultAgent(TemplateAgent):
                 return
 
             if status == "finished":
-                self._register_embedding_search_tool(
-                    payload.get("vector_store_instance")
-                )
+                self._register_embedding_search_tool(payload.get("vector_store_instance"))
 
             text = await self._build_embedding_direct_output(
                 status=status,
@@ -624,15 +881,75 @@ class DefaultAgent(TemplateAgent):
         async for item in self._stream_request(request):
             yield item
 
+    def restore_history(self, messages: list[dict[str, Any]]) -> None:
+        """Restore persisted conversation history into the session state.
+
+        Parameters
+        ----------
+        messages : list[dict[str, Any]]
+            Persisted chat messages.
+        """
+
+        restored: list[BaseMessage] = []
+        for message in messages:
+            role = message.get("role")
+            content = str(message.get("content", ""))
+            if role == "user":
+                restored.append(HumanMessage(content=content))
+            elif role == "assistant":
+                restored.append(AIMessage(content=content))
+
+        self._session.messages = restored
+        self._set_system_prompt(self._build_turn_state())
+
+    def get_chat_history(self, with_system: bool = False) -> str | None:
+        """Render plain-text chat history."""
+
+        try:
+            return _format_chat_history(self._session.messages, with_system=with_system)
+        except Exception as exc:
+            logger.warning("Failed to build chat history: %s", exc)
+            return None
+
+    def clone(self) -> Agent:
+        """Clone the agent with a fresh session.
+
+        Returns
+        -------
+        Agent
+            Session-safe cloned agent.
+        """
+
+        return type(self)(
+            model=self.model,
+            system_prompt=self.system_prompt,
+            voice_names=list(self.voice_names),
+            emotions=list(self.emotions),
+            tools=self._tool_provider.get_tool_specs(),
+        )
+
+    def add_tools(self, tools: list[BaseTool | Callable[[], BaseTool]]) -> None:
+        """Attach additional tools to the agent.
+
+        Parameters
+        ----------
+        tools : list[BaseTool | Callable[[], BaseTool]]
+            Tool instances or factories.
+        """
+
+        if not tools:
+            return
+        self._tool_provider.add_tools(tools)
+
     def _register_embedding_search_tool(self, vector_store_instance: Any) -> None:
         """Register the local-search tool for uploaded documents when ready."""
 
         if vector_store_instance is None:
             return
-        dynamic_tools = self.runtime.session.metadata.setdefault("dynamic_tools", {})
+        dynamic_tools = self._session.metadata.setdefault("dynamic_tools", {})
         if not isinstance(dynamic_tools, dict):
             dynamic_tools = {}
-            self.runtime.session.metadata["dynamic_tools"] = dynamic_tools
+            self._session.metadata["dynamic_tools"] = dynamic_tools
         if LOCAL_SEARCH_TOOL in dynamic_tools:
             return
         try:
@@ -650,7 +967,7 @@ class DefaultAgent(TemplateAgent):
     ) -> str | None:
         """Build one direct-output utterance for embedding lifecycle updates."""
 
-        history = _format_chat_history(self.runtime.session.messages) or ""
+        history = _format_chat_history(self._session.messages) or ""
         if status == "processing":
             prompt_history = [
                 SystemMessage(content=_EMBEDDING_PROCESSING_PROMPT),
@@ -668,7 +985,7 @@ class DefaultAgent(TemplateAgent):
         try:
             response = await self.model.ainvoke(prompt_history)
             content = getattr(response, "content", "")
-            text = self.runtime.scenario.output_policy.filter_text(str(content or ""))
+            text = self._filter_text(str(content or ""))
             if text:
                 return text
         except Exception as exc:
@@ -684,7 +1001,7 @@ class DefaultAgent(TemplateAgent):
     ) -> str:
         """Return a deterministic fallback embedding utterance."""
 
-        history = _format_chat_history(self.runtime.session.messages) or ""
+        history = _format_chat_history(self._session.messages) or ""
         prefers_chinese = bool(re.search(r"[\u4e00-\u9fff]", history + doc_text))
         if status == "processing":
             return (
