+++ src/xtalk/llm_agent/interfaces.py
@@ -1,9 +1,11 @@
 import asyncio
 from abc import ABC, abstractmethod
 from typing import Iterable, TypedDict, AsyncIterator, Callable, Any, Union
+
 from langchain_core.messages import ToolCall
 from langchain_core.tools import BaseTool
-from ..tools.utils import ToolCallResultPayload
+
+from .tools.utils import ToolCallResultPayload
 
 
 class AgentContext(TypedDict):
