+++ src/xtalk/serving/module_types.py
@@ -6,7 +6,6 @@ from .modules.latency_manager import LatencyManager
 from .modules.llm_agent_context_manager import LLMAgentContextManager
 from .modules.llm_agent_generation_manager import LLMAgentConsumptionManager
 from .modules.speaker_manager import SpeakerManager
-from .modules.thought_manager import ThoughtManager
 from .modules.tts_manager import TTSManager
 from .modules.turn_taking_manager import TurnTakingManager
 from .modules.vad_manager import VADManager
@@ -20,7 +19,6 @@ __all__ = [
     "LLMAgentContextManager",
     "LLMAgentConsumptionManager",
     "SpeakerManager",
-    "ThoughtManager",
     "TTSManager",
     "TurnTakingManager",
     "VADManager",
